pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gcp-actions"
version = "0.1.15"
description = "Reusable Google Cloud helper functions"
authors = [{ name = "Stanislav", email = "bike4cats@gmail.com" }]
requires-python = ">=3.8"

# CORE DEPENDENCIES (Keep this list absolutely minimal)
# These are installed for EVERY service that imports gcp-actions.
dependencies = [
    # Core auth libraries are needed by all other gcp libraries
    "google-auth",
    "google-api-core",
    "google-cloud-logging",
    "python-dotenv",
    "uuid"

]
# -----------------------------------------------------------------
# OPTIONAL DEPENDENCIES
# -----------------------------------------------------------------
# These are only installed if explicitly requested.
[project.optional-dependencies]
storage = ["google-cloud-storage"]
pubsub = ["google-cloud-pubsub"]
firestore = ["google-cloud-firestore"]
secretmanager = ["google-cloud-secret-manager"]

# Installs all dependencies for your data processing worker
processing-worker = [
    "gcp-actions[storage]",
    "gcp-actions[pubsub]",
    "gcp-actions[firestore]",
    "gcp-actions[secretmanager]",
]

[tool.setuptools.packages.find]
where = ["."]